requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.1
django-admin-rangefilter==0.8.4
django-cors-headers==3.11.0
django-encrypted-field==1.0.4
django-environ==0.8.1
django-parler==2.3
django-regex-field==2.0.2
django-rosetta==0.9.8
django-templated-mail==1.1.1
django-user-agents==0.4.0
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
Faker==11.3.0
flake8==4.0.1
gunicorn==19.7.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.2.0
packaging==21.3
passlib==1.7.4
Pillow==9.0.1
pluggy==1.0.0
polib==1.1.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.15.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==7.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
ua-parser==0.15.0
uritemplate==4.1.1
urllib3==1.26.8
user-agents==2.2.0
